move Bench code into separate module
