+++ api/src/wrangler.rs
@@ -11,8 +11,9 @@ use fvm_shared::econ::TokenAmount;
 use fvm_shared::message::Message;
 use fvm_shared::{ActorID, MethodNum, BLOCK_GAS_LIMIT};
 
-pub use crate::{ActorState, Bench, ExecutionResult};
+pub use crate::bench::{ActorState, Bench, ExecutionResult};
 
+/// High level wrapper of a Bench for convenience
 pub struct ExecutionWrangler {
     bench: RefCell<Box<dyn Bench>>,
     store: Box<dyn Blockstore>,
